+++ Tests/AdvancedOperationTests/Core/AsynchronousOperationTests.swift
@@ -128,6 +128,7 @@ final class AsynchronousOperationTests: XCTestCase {
   }
 
   func testMultipleStart() {
+    // TODO: at the moment multiple starts are allowed
     let operation = SleepyAsyncOperation()
     let expectation1 = XCTKVOExpectation(keyPath: #keyPath(Operation.isFinished), object: operation, expectedValue: true)
     let expectation2 = expectation(description: "\(#function)\(#line)")
@@ -372,20 +373,6 @@ final class AsynchronousOperationTests: XCTestCase {
     gate.addDependencies(operation1, operation2)
     operation3.addDependencies(gate)
 
-    if #available(iOS 12.0, iOSApplicationExtension 12.0, tvOS 12.0, watchOS 5.0, macOS 10.14, OSXApplicationExtension 10.14, *){
-      operation0.installLogger(log: Log.default, signpost: Log.signpost, poi: Log.poi)
-      operation1.installLogger(log: Log.default, signpost: Log.signpost, poi: Log.poi)
-      operation2.installLogger(log: Log.default, signpost: Log.signpost, poi: Log.poi)
-      operation3.installLogger(log: Log.default, signpost: Log.signpost, poi: Log.poi)
-      gate.installLogger(log: Log.default, signpost: Log.signpost, poi: Log.poi)
-    } else {
-      operation0.installLogger(log: Log.default, signpost: Log.signpost, poi: .disabled)
-      operation1.installLogger(log: Log.default, signpost: Log.signpost, poi: .disabled)
-      operation2.installLogger(log: Log.default, signpost: Log.signpost, poi: .disabled)
-      operation3.installLogger(log: Log.default, signpost: Log.signpost, poi: .disabled)
-      gate.installLogger(log: Log.default, signpost: Log.signpost, poi: .disabled)
-    }
-
     operation0.name = "op0"
     operation1.name = "op1"
     operation2.name = "op2"
@@ -399,7 +386,7 @@ final class AsynchronousOperationTests: XCTestCase {
   }
 }
 
-final class GateOperation: Operation, LoggableOperation {
+final class GateOperation: Operation {
   private let block: (GateOperation) -> Void
 
   init(block: @escaping (GateOperation) -> Void) {