+++ Tests/AdvancedOperationTests/Operations/AsynchronousBlockOperationTests.swift
@@ -31,6 +31,7 @@ final class AsynchronousBlockOperationTests: XCTestCase {
     #endif
   }
 
+  // TODO
 //  func testCancelledExecution() {
 //    let operation = AsynchronousBlockOperation { complete in
 //      DispatchQueue(label: "\(identifier).\(#function)", attributes: .concurrent).asyncAfter(deadline: .now() + 2) {
@@ -63,6 +64,7 @@ final class AsynchronousBlockOperationTests: XCTestCase {
     XCTAssertTrue(operation.isCancelled)
   }
 
+    // TODO
 //  func testEarlyBailOut() {
 //    let operation = AsynchronousBlockOperation { $0() }
 //    let expectation1 = XCTKVOExpectation(keyPath: #keyPath(Operation.isFinished), object: operation, expectedValue: true)
@@ -94,10 +96,7 @@ final class AsynchronousBlockOperationTests: XCTestCase {
         complete()
       }
     }
-    operation.addCompletionBlock {
-      expectation2.fulfill()
-    }
-    operation.installLogger()
+    operation.addCompletionBlock { expectation2.fulfill() }
     operation.start()
     wait(for: [expectation1, expectation2], timeout: 10, enforceOrder: true)
   }